+++ plugins/comparative-adapters/src/MCScanAnchorsAdapter/MCScanAnchorsAdapter.ts
@@ -2,11 +2,14 @@ import {
   BaseFeatureDataAdapter,
   BaseOptions,
 } from '@jbrowse/core/data_adapters/BaseAdapter'
-import { Region } from '@jbrowse/core/util/types'
 import { openLocation } from '@jbrowse/core/util/io'
-import { doesIntersect2 } from '@jbrowse/core/util'
+import {
+  doesIntersect2,
+  SimpleFeature,
+  Feature,
+  Region,
+} from '@jbrowse/core/util'
 import { ObservableCreate } from '@jbrowse/core/util/rxjs'
-import SimpleFeature, { Feature } from '@jbrowse/core/util/simpleFeature'
 import { readFile, parseBed } from '../util'
 
 interface BareFeature {
@@ -102,15 +105,16 @@ export default class MCScanAnchorsAdapter extends BaseFeatureDataAdapter {
                 ...f1,
                 uniqueId: `${index}-${rowNum}`,
                 syntenyId: rowNum,
-                // note: strand would be -1 if the two features are on opposite strands,
-                // indicating inverted alignment
+
+                // note: strand would be -1 if the two features are on opposite
+                // strands, indicating inverted alignment
                 strand: f1.strand * f2.strand,
                 assemblyName: assemblyNames[+!flip],
                 score,
                 mate: {
                   ...f2,
                   assemblyName: assemblyNames[+flip],
-                } as BareFeature,
+                },
               }),
             )
           }
