+++ plugins/linear-comparative-view/src/LinearReadVsRef/LinearReadVsRef.tsx
@@ -2,70 +2,31 @@ import React, { useEffect, useState } from 'react'
 import {
   Button,
   CircularProgress,
-  Checkbox,
   Dialog,
   DialogActions,
   DialogContent,
   DialogTitle,
-  FormControlLabel,
   IconButton,
   TextField,
   Typography,
 } from '@mui/material'
 import { makeStyles } from 'tss-react/mui'
-import { MismatchParser } from '@jbrowse/plugin-alignments'
 import { getConf } from '@jbrowse/core/configuration'
 import { getRpcSessionId } from '@jbrowse/core/util/tracks'
 import { getSession, getContainingView, Feature } from '@jbrowse/core/util'
 
+// locals
+import {
+  getClip,
+  getLength,
+  getLengthOnRef,
+  getLengthSansClipping,
+  getTag,
+  mergeIntervals,
+} from './util'
+
 // icons
 import CloseIcon from '@mui/icons-material/Close'
-
-const { parseCigar } = MismatchParser
-
-function getLengthOnRef(cigar: string) {
-  const cigarOps = parseCigar(cigar)
-  let lengthOnRef = 0
-  for (let i = 0; i < cigarOps.length; i += 2) {
-    const len = +cigarOps[i]
-    const op = cigarOps[i + 1]
-    if (op !== 'H' && op !== 'S' && op !== 'I') {
-      lengthOnRef += len
-    }
-  }
-  return lengthOnRef
-}
-
-function getLength(cigar: string) {
-  const cigarOps = parseCigar(cigar)
-  let length = 0
-  for (let i = 0; i < cigarOps.length; i += 2) {
-    const len = +cigarOps[i]
-    const op = cigarOps[i + 1]
-    if (op !== 'D' && op !== 'N') {
-      length += len
-    }
-  }
-  return length
-}
-
-function getLengthSansClipping(cigar: string) {
-  const cigarOps = parseCigar(cigar)
-  let length = 0
-  for (let i = 0; i < cigarOps.length; i += 2) {
-    const len = +cigarOps[i]
-    const op = cigarOps[i + 1]
-    if (op !== 'H' && op !== 'S' && op !== 'D' && op !== 'N') {
-      length += len
-    }
-  }
-  return length
-}
-function getClip(cigar: string, strand: number) {
-  return strand === -1
-    ? +(cigar.match(/(\d+)[SH]$/) || [])[1] || 0
-    : +(cigar.match(/^(\d+)([SH])/) || [])[1] || 0
-}
 
 interface ReducedFeature {
   refName: string
@@ -96,51 +57,6 @@ const useStyles = makeStyles()(theme => ({
     color: theme.palette.grey[500],
   },
 }))
-
-function getTag(f: Feature, tag: string) {
-  const tags = f.get('tags')
-  return tags ? tags[tag] : f.get(tag)
-}
-
-function mergeIntervals<T extends { start: number; end: number }>(
-  intervals: T[],
-  w = 5000,
-) {
-  // test if there are at least 2 intervals
-  if (intervals.length <= 1) {
-    return intervals
-  }
-
-  const stack = []
-  let top = null
-
-  // sort the intervals based on their start values
-  intervals = intervals.sort((a, b) => a.start - b.start)
-
-  // push the 1st interval into the stack
-  stack.push(intervals[0])
-
-  // start from the next interval and merge if needed
-  for (let i = 1; i < intervals.length; i++) {
-    // get the top element
-    top = stack[stack.length - 1]
-
-    // if the current interval doesn't overlap with the
-    // stack top element, push it to the stack
-    if (top.end + w < intervals[i].start - w) {
-      stack.push(intervals[i])
-    }
-    // otherwise update the end value of the top element
-    // if end of current interval is higher
-    else if (top.end < intervals[i].end) {
-      top.end = Math.max(top.end, intervals[i].end)
-      stack.pop()
-      stack.push(top)
-    }
-  }
-
-  return stack
-}
 
 interface BasicFeature {
   end: number
@@ -182,7 +98,6 @@ export default function ReadVsRefDialog({
   const [windowSizeText, setWindowSize] = useState('0')
   const [error, setError] = useState<unknown>()
   const [primaryFeature, setPrimaryFeature] = useState<Feature>()
-  const [qualTrack, setQualTrack] = useState(false)
   const windowSize = +windowSizeText
 
   // we need to fetch the primary alignment if the selected feature is 2048.
@@ -248,7 +163,6 @@ export default function ReadVsRefDialog({
       const view = getContainingView(track)
       const cigar = feature.get('CIGAR') as string
       const flags = feature.get('flags') as number
-      const qual = feature.get('qual') as string
       const origStrand = feature.get('strand') as number
       const SA = (getTag(feature, 'SA') as string) || ''
       const readName = feature.get('name') as string
@@ -425,39 +339,6 @@ export default function ReadVsRefDialog({
                   },
                 ],
               },
-
-              ...(qualTrack
-                ? [
-                    {
-                      id: `${Math.random()}`,
-                      type: 'QuantitativeTrack',
-                      configuration: {
-                        trackId: 'qualTrack',
-                        assemblyNames: [readAssembly],
-                        name: 'Read quality',
-                        type: 'QuantitativeTrack',
-                        adapter: {
-                          type: 'FromConfigAdapter',
-                          noAssemblyManager: true,
-                          features: qual.split(' ').map((score, index) => ({
-                            start: index,
-                            end: index + 1,
-                            refName: readName,
-                            score: +score,
-                            uniqueId: `feat_${index}`,
-                          })),
-                        },
-                      },
-                      displays: [
-                        {
-                          id: `${Math.random()}`,
-                          type: 'LinearWiggleDisplay',
-                          height: 100,
-                        },
-                      ],
-                    },
-                  ]
-                : []),
             ],
           },
         ],
@@ -528,15 +409,6 @@ export default function ReadVsRefDialog({
               value={windowSize}
               onChange={event => setWindowSize(event.target.value)}
               label="Set window size"
-            />
-            <FormControlLabel
-              control={
-                <Checkbox
-                  checked={qualTrack}
-                  onChange={() => setQualTrack(val => !val)}
-                />
-              }
-              label="Show qual track"
             />
           </div>
         )}
